+++ cloud-minecraft/cloud-bukkit/src/main/java/cloud/commandframework/bukkit/parsers/selector/SelectorParseException.java
@@ -69,7 +69,7 @@ public final class SelectorParseException extends ParserException {
      * @return String value
      */
     public @NonNull String getInput() {
-        return input;
+        return this.input;
     }
 
     /**