+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - JDA Role argument parser
+
+### Changed
+ - Use Command instead of TabCompleteEvent on Bukkit
+ - Add isolation system to JDA user argument
+
+### Deprecated
+ - Deprecated old JDA UserParser that did not take an isolation parameter 
+ 
+### Fixed
+ - Tuple implementations now do not throw an error when using the toArray method
+ - Argument parser for `long` types was not registered
+
 ## [1.4.0] - 2021-01-16
 
 ### Added